move _is_authenticated, _get_ldap_agent, logged_in_user, nfp_for_country orgs_in_country to logic_common _get_ldap_agent default bind=True remove edit user permission for nfps removed unused code for auto-disabling users
