+++ eea/ldapadmin/roles_editor.py
@@ -34,6 +34,8 @@ from eea.ldapadmin.ui_common import (CommonTemplateLogic,
                                      NaayaViewPageTemplateFile,
                                      TemplateRenderer, get_role_name,
                                      roles_list_to_text)
+from eea.ldapadmin.logic_common import logged_in_user, _is_authenticated
+from eea.ldapadmin.logic_common import _get_ldap_agent
 from OFS.Folder import Folder
 from persistent.mapping import PersistentMapping
 from Products.Five.browser import BrowserView
@@ -72,11 +74,6 @@ def manage_add_roles_editor(parent, tool_id, REQUEST=None):
         REQUEST.RESPONSE.redirect(parent.absolute_url() + '/manage_workspace')
 
 
-def _is_authenticated(request):
-    ''' check if user is authenticated '''
-    return 'Authenticated' in request.AUTHENTICATED_USER.getRoles()
-
-
 def _role_parents(role_id):
     ''' get role parents '''
     if role_id is None:
@@ -88,19 +85,6 @@ def _role_parents(role_id):
         parents.append(role_id)
 
     return reversed(parents)
-
-
-def logged_in_user(request):
-    ''' return the id of the authenticated user '''
-    user_id = ''
-
-    if _is_authenticated(request):
-        user = request.get('AUTHENTICATED_USER', '')
-
-        if user:
-            user_id = user.getId()
-
-    return user_id
 
 
 def filter_roles(agent, pattern):
@@ -318,13 +302,6 @@ class RolesEditor(Folder):
         self._config.update(ldap_config.read_form(REQUEST.form, edit=True))
         REQUEST.RESPONSE.redirect(self.absolute_url() + '/manage_edit')
 
-    def _get_ldap_agent(self, bind=True):
-        ''' get the ldap agent '''
-        agent = ldap_config.ldap_agent_with_config(self._config, bind)
-        agent._author = logged_in_user(self.REQUEST)
-
-        return agent
-
     def _predefined_filters(self):
         ''' return predefined filters '''
         return sorted(self.objectValues([query.Query.meta_type]),
@@ -346,7 +323,7 @@ class RolesEditor(Folder):
     def index_html(self, REQUEST):
         """ view """
         role_id = REQUEST.form.get('role_id', None)
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         try:
             role_info = agent.role_info(role_id)
         except usersdb.RoleNotFound:
@@ -415,7 +392,7 @@ class RolesEditor(Folder):
     def table_problematic_roles(self):
         """ A table of all problematic roles
         """
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         roles = []
 
         for role_cn, role_info in agent.all_roles():
@@ -440,7 +417,7 @@ class RolesEditor(Folder):
         breadcrumbs = [('Search', search_url)]
 
         if pattern:
-            agent = self._get_ldap_agent()
+            agent = _get_ldap_agent(self)
             results_html = filter_result_html(agent, pattern,
                                               self._render_template)
             options['results_html'] = results_html
@@ -469,7 +446,7 @@ class RolesEditor(Folder):
             return "You must be logged in to access this page.\n"
 
         pattern = REQUEST.form.get('pattern', '')
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
 
         output_file = StringIO()
         csv_file = csv.writer(output_file)
@@ -608,21 +585,21 @@ class RolesEditor(Folder):
 
     def _merge_roles(self, roles):
         ''' merge roles '''
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
 
         for role_source, role_destination in roles.items():
             agent.merge_roles(role_source, role_destination)
 
     def _prefill_roles(self, roles):
         ''' prefill roles '''
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
 
         for role_destination, role_source in roles.items():
             agent.prefill_roles_from(role_destination, role_source)
 
     def _create_roles(self, roles):
         ''' create roles '''
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
 
         problems = []
 
@@ -646,7 +623,7 @@ class RolesEditor(Folder):
     def _rename_roles(self, roles):
         """ This is actually just changing their description
         """
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
 
         problems = []
 
@@ -693,7 +670,7 @@ class RolesEditor(Folder):
 
             return False
 
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         role_info = agent.role_info(role_id)
 
         return str.encode(agent._user_dn(user.getId())) in role_info['owner']
@@ -721,7 +698,7 @@ class RolesEditor(Folder):
             # top role - can_edit_roles check was sufficient for granting
 
             return False
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         role_info = agent.role_info(role_id)
 
         if agent._user_dn(user.getId()) not in role_info['owner']:
@@ -798,7 +775,7 @@ class RolesEditor(Folder):
             return self.create_role_html(REQUEST)
 
         user_id = logged_in_user(REQUEST)
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         slug = REQUEST.form['slug']
         description = REQUEST.form['description']
         parent_role_id = REQUEST.form.get('parent_role_id', '') or None
@@ -858,7 +835,7 @@ class RolesEditor(Folder):
             raise Unauthorized(("You are not allowed to delete role %s. "
                                 "Owners can only delete empty roles")
                                % role_id)
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
 
         to_remove = list(map(agent._role_id, agent._sub_roles(role_id)))
         options = {
@@ -882,7 +859,7 @@ class RolesEditor(Folder):
                                 "Owners can only delete empty roles")
                                % role_id)
         logged_in = logged_in_user(REQUEST)
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         # first remove users from role
         with agent.new_action():
             for user_id in agent.members_in_role_and_subroles(role_id)[
@@ -915,7 +892,7 @@ class RolesEditor(Folder):
         }
 
         if search_name:
-            agent = self._get_ldap_agent()
+            agent = _get_ldap_agent(self)
             found_users = agent.search_user(search_name)
             active_users = [user for user in found_users
                             if user.get('status') != 'disabled']
@@ -939,7 +916,7 @@ class RolesEditor(Folder):
             raise Unauthorized("You are not allowed to manage members in %s" %
                                role_id)
         user_id = REQUEST.form['user_id']
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         with agent.new_action():
             role_id_list = agent.add_to_role(role_id, 'user', user_id)
         roles_msg = roles_list_to_text(agent, role_id_list)
@@ -959,7 +936,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage members in %s" %
                                role_id)
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         options = {
             'role_id': role_id,
             'role_members': role_members(agent, role_id),
@@ -974,7 +951,7 @@ class RolesEditor(Folder):
 
     def remove_members(self, REQUEST):
         """ Remove user several members from a role """
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         role_id = REQUEST.form['role_id']
 
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
@@ -1009,7 +986,7 @@ class RolesEditor(Folder):
             raise Unauthorized("You are not allowed to manage members in %s" %
                                role_id)
         user_id = REQUEST.form['user_id']
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         user_roles = agent.list_member_roles('user', user_id)
         options = {
             'role_id': role_id,
@@ -1026,7 +1003,7 @@ class RolesEditor(Folder):
         this action. Called by remove_user_from_role and delete_role
 
         """
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         with agent.new_action():
             role_id_list = agent.remove_from_role(role_id, 'user', user_id)
         log.info("%s REMOVED USER %r FROM ROLE(S) %r",
@@ -1066,13 +1043,15 @@ class RolesEditor(Folder):
             'user_id': user_id,
         }
 
-        if search_name:
-            agent = self._get_ldap_agent()
-            options['search_results'] = agent.search_user(search_name)
+        if search_name or user_id:
+            agent = _get_ldap_agent(self)
 
-        if user_id is not None:
-            agent = self._get_ldap_agent()
-            options['user_roles'] = agent.list_member_roles('user', user_id)
+            if search_name:
+                options['search_results'] = agent.search_user(search_name)
+
+            if user_id is not None:
+                options['user_roles'] = agent.list_member_roles('user',
+                                                                user_id)
 
         return self._render_template('zpt/roles_search_users.zpt', **options)
 
@@ -1100,7 +1079,7 @@ class RolesEditor(Folder):
         if subroles:
             header = ('Subrole', ) + header
 
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         try:
             agent.role_info(role_id)
         except usersdb.RoleNotFound:
@@ -1135,7 +1114,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage members in %s" %
                                role_id)
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         options = {'role_id': role_id,
                    'role_owners': {}}
         user_id_list = REQUEST.form.get('user_id_list', [])
@@ -1208,7 +1187,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage senders in %s" %
                                role_id)
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         data = agent.mail_group_info(role_id)
         data['emails'] = '\n'.join([e for e in data['permittedSender']
                                     if '@' in e])
@@ -1235,7 +1214,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage senders in %s" %
                                role_id)
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         msgs = IStatusMessage(REQUEST)
 
         if REQUEST.REQUEST_METHOD == 'POST':
@@ -1290,7 +1269,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage senders in %s" %
                                role_id)
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         msgs = IStatusMessage(REQUEST)
 
         if REQUEST.REQUEST_METHOD == 'POST':
@@ -1346,7 +1325,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage leaders in %s" %
                                role_id)
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         leaders, alternates = agent.role_leaders(role_id)
         members = {}
 
@@ -1377,7 +1356,7 @@ class RolesEditor(Folder):
         if not self.can_edit_members(role_id, REQUEST.AUTHENTICATED_USER):
             raise Unauthorized("You are not allowed to manage leaders in %s" %
                                role_id)
-        agent = self._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self)
         crt_leaders, crt_alternates = agent.role_leaders(role_id)
         leader = REQUEST.form.get('leader')
         alternates = REQUEST.form.get('alternate_list', [])
@@ -1428,7 +1407,7 @@ class RolesEditor(Folder):
             raise Unauthorized("You are not allowed to manage members in %s" %
                                role_id)
 
-        agent = self._get_ldap_agent()
+        agent = _get_ldap_agent(self)
         try:
             role_info = agent.role_info(role_id)
         except usersdb.RoleNotFound:
@@ -1471,7 +1450,7 @@ class RolesEditor(Folder):
 
         if REQUEST.REQUEST_METHOD == 'POST':
             description = REQUEST.form.get('role_name')
-            agent = self._get_ldap_agent(bind=True)
+            agent = _get_ldap_agent(self)
             try:
                 with agent.new_action():
                     agent.set_role_description(role_id, description)
@@ -1541,7 +1520,7 @@ class RolesStatistics(BrowserView):
 
     def __call__(self):
 
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         this_role_id = self.request.form.get('role_id')
         assert this_role_id
 
@@ -1576,7 +1555,7 @@ class ExtendedManagementEditor(BrowserView):
 
     def handle_enable_extended_management(self):
         ''' set extended management status '''
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         role_id = self.request.form.get('role_id')
         assert role_id
         is_extended = self.request.form.get('is_extended') == 'on'
@@ -1587,7 +1566,7 @@ class ExtendedManagementEditor(BrowserView):
 
     def handle_empty_branch(self):
         ''' handle empty branch '''
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         role_id = self.request.form.get('role_id')
         assert role_id
 
@@ -1631,7 +1610,7 @@ class ExtendedManagementEditor(BrowserView):
 
     def view(self):
         ''' render statistics index '''
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         this_role_id = self.request.form.get('role_id')
         assert this_role_id
 
@@ -1681,7 +1660,7 @@ class IsExtendedEnabled(BrowserView):
     """
 
     def __call__(self, role_id):
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
 
         if not role_id:
             return False
@@ -1741,7 +1720,7 @@ class EditMembersOfOneRole(BrowserView):
 
     def view(self):
         ''' edit members of one role - view '''
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         this_role_id = self.request.form.get('role_id')
         assert this_role_id
 
@@ -1791,7 +1770,7 @@ class EditMembersOfOneRole(BrowserView):
 
     def processForm(self):
         ''' form processing '''
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         role_id = self.request.form.get('role_id')
         assert role_id
 
@@ -1850,7 +1829,7 @@ class EditRolesOfOneMember(BrowserView):
     def view(self):
         ''' main view '''
         selected_member = self.request.form.get('member')
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
 
         this_role_id = self.request.form.get('role_id')
 
@@ -1906,7 +1885,7 @@ class EditRolesOfOneMember(BrowserView):
 
     def processForm(self):
         ''' form processing '''
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         role_id = self.request.form.get('role_id')
         assert role_id
 
@@ -1959,7 +1938,7 @@ class ExportExcel(BrowserView):
     ''' Export to Excel '''
 
     def __call__(self):
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         this_role_id = self.request.form.get('role_id')
         assert this_role_id
 
@@ -2033,7 +2012,7 @@ class SearchEionet(BrowserView):
     ''' ldap search vor javascript frontend '''
     def __call__(self):
         ldap_filter = self.request.form.get('filter')
-        agent = self.context._get_ldap_agent(bind=True)
+        agent = _get_ldap_agent(self.context)
         encres = json.dumps(agent.search_user(ldap_filter))
         self.request.response.setHeader("Content-Type", 'application/json')
         self.request.response.setHeader("Content-Length", str(len(encres)))
