+++ eea/ldapadmin/ui_common.py
@@ -1,11 +1,10 @@
 ''' common ui methods '''
-import six
 from zope.component import getMultiAdapter
 from Acquisition import Implicit
 from eea.ldapadmin.constants import NETWORK_NAME
 from eea.ldapadmin import roles_leaders
-from eea.ldapadmin.countries import get_country, get_country_options
-from eea.ldapadmin.logic_common import _get_user_id, _is_authenticated
+from eea.ldapadmin.countries import get_country
+from eea.ldapadmin.logic_common import logged_in_user, _is_authenticated
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PageTemplates.PageTemplateFile import PageTemplateFile
 from z3c.pt.pagetemplate import PageTemplateFile as ChameleonTemplate
@@ -53,49 +52,6 @@ def extend_crumbs(crumbs_html, editor_url, extra_crumbs):
     last_crumb.text = last_crumb_text
 
     return tostring(crumbs)
-
-
-def orgs_in_country(context, country):
-    """ return a dict of organisations in countrys """
-    agent = context._get_ldap_agent(secondary=True)
-    orgs_by_id = agent.all_organisations()
-    countries = dict(get_country_options(country=country))
-    orgs = {}
-
-    for org_id, info in six.iteritems(orgs_by_id):
-        country_info = countries.get(info['country'])
-
-        if country_info:
-            orgs[org_id] = info
-
-    return orgs
-
-
-def nfp_for_country(context):
-    """ Return country code for which the current user has NFP role
-        or None otherwise"""
-    user_id = context.REQUEST.AUTHENTICATED_USER.getId()
-
-    if user_id:
-        ldap_groups = get_ldap_user_groups(context, user_id)
-
-        for group in ldap_groups:
-            if ('eionet-nfp-mc-' in group[0] or
-                'eionet-nfp-cc-' in group[0] or
-                    'eionet-nfp-oc-' in group[0]):
-
-                return group[0].rsplit('-', 1)[-1]
-    return None
-
-
-def get_ldap_user_groups(context, user_id):
-    """ return the ldap roles the user is member of """
-    agent = context._get_ldap_agent(bind=True, secondary=True)
-    ldap_roles = sorted(agent.member_roles_info('user',
-                                                user_id,
-                                                ('description',)))
-
-    return ldap_roles
 
 
 # pylint: disable=dangerous-default-value
@@ -225,7 +181,7 @@ class CommonTemplateLogic(object):
 
     def user_id(self):
         ''' return the user id '''
-        return _get_user_id(self._get_request())
+        return logged_in_user(self._get_request())
 
     def readonly_alert(self):
         ''' return a readonly alert message '''
