+++ eea/ldapadmin/views/changelog.py
@@ -5,6 +5,7 @@ from zope.interface import Attribute, Interface, implementer
 from DateTime.DateTime import DateTime
 from eea.usersdb import factories
 from eea.usersdb.db_agent import UserNotFound
+from eea.ldapadmin.logic_common import _get_ldap_agent
 from Products.Five import BrowserView
 
 
@@ -39,7 +40,7 @@ class BaseActionDetails(BrowserView):
             return entry['author']
 
         try:
-            user_info = self.base._get_ldap_agent().user_info(entry['author'])
+            user_info = _get_ldap_agent(self.base).user_info(entry['author'])
         except UserNotFound:
             return entry['author']
 
@@ -51,7 +52,7 @@ class BaseActionDetails(BrowserView):
             return ''
 
         try:
-            agent = self.base._get_ldap_agent()
+            agent = _get_ldap_agent(self.base)
             user_info = agent.user_info(entry['author'])
         except UserNotFound:
             return ''
@@ -149,7 +150,7 @@ class OrganisationChangelog(BrowserView):
     def entries(self):
         ''' get changelog entries '''
         org_id = self.request.form.get('id').encode()
-        agent = self.context._get_ldap_agent()
+        agent = _get_ldap_agent(self.context)
         org_dn = agent._org_dn(org_id)
 
         log_entries = list(reversed(agent._get_metadata(org_dn)))
