+++ eea/ldapadmin/zpt/nfp_nrc/extranet_reporters.zpt
@@ -97,7 +97,7 @@
                   ).append('&nbsp;');
                   org.append(
                     $('<a>')
-                      .attr({href: 'https://www.eionet.europa.eu/ldap-organisations/organisation?id='
+                      .attr({href: '/ldap-organisations/organisation?id='
                                     + user_info['ldap_org']['id']})
                       .text(user_info['ldap_org']['name'])
                   ).append('&nbsp;');
