--- eea/ldapadmin/zpt/users/email_auto_disabled.zpt
@@ -1,8 +1,0 @@
-Hello <span tal:replace="options/full_name"></span>,
-
-This is an automated email from <span tal:replace="options/site_title"></span>!
-
-Your account with user ID `<span tal:replace="options/username"></span>` has
-been inactive for a long period of time and has been disabled.
-
-<span tal:replace="options/site_title"></span>
