--- eea/ldapadmin/zpt/users/email_auto_predisabled.zpt
@@ -1,10 +1,0 @@
-Hello <span tal:replace="options/full_name"></span>,
-
-This is an automated email from <span tal:replace="options/site_title"></span>!
-
-Your account with user ID `<span tal:replace="options/username"></span>` has
-been inactive for a long period of time. You have 30 days to login with your
-account, after which the user will be automatically disabled.
-
-<span tal:replace="options/site_title"></span>
-
