--- eea/ldapadmin/zpt/users/email_report_autodisable.zpt
@@ -1,13 +1,0 @@
-Hello Helpdesk,
-
-This is an automated email from <span tal:replace="options/site_title"></span>!
-
-Following users has been sent messages that their account will be disabled if they don't login in a month:
-
-<span tal:repeat="user options/users_predisable" tal:omit-tag=""><span tal:replace="user/id"></span></span>
-
-The following users have been automatically disabled, they haven't logged in <span tal:replace="options/days"></span> days:
-
-<span tal:repeat="user options/users_disable" tal:omit-tag=""><span tal:replace="user/id"></span></span>
-
-<span tal:replace="options/site_title"></span>
